♻️ [Refactoring]: #456 lexer/hex_string.rs テストモジュールを 9 ファイルに物理分割
568 行に肥大化していた hex_string.rs 末尾の `mod tests` (45 テスト関数) を
カテゴリ単位で 9 つのサブモジュールへ物理分割し、本体ファイルを 112 行に縮小。
プロダクトコードと既存テスト関数の本体は一切変更せず、配置のみを移動。

- basic / byte_preservation / case / error / guard /
  padding / robustness / rollback / whitespace に分類
- 本体側は `#[cfg(test)] mod tests;` の宣言のみに置換
- `lexer/hex_string/tests.rs` に各サブモジュールの mod 宣言を集約

Refs #456
