+++ rust/crates/core/src/lexer/hex_string/tests.rs
@@ -1,0 +1,9 @@
+mod basic;
+mod byte_preservation;
+mod case;
+mod error;
+mod guard;
+mod padding;
+mod robustness;
+mod rollback;
+mod whitespace;
